Move plotting function into report.py
